added limit by memory to lrucache

+++ lru/lrucache.inl
@@ -1,36 +1,36 @@
 #include "lrucache.hpp"
 
-template <typename K, typename T, size_t mem, typename lookup_t>
+template <typename K, typename T, ull mem, typename lookup_t>
 auto LruCache<K, T, mem, lookup_t>::cbegin() const
 {
     return items.cbegin();
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 auto LruCache<K, T, mem, lookup_t>::cend() const
 {
     return items.cend();
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 size_t LruCache<K, T, mem, lookup_t>::size() const
 {
     return items.size();
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
-size_t LruCache<K, T, mem, lookup_t>::max_items() const
+template <typename K, typename T, ull mem, typename lookup_t> 
+ull LruCache<K, T, mem, lookup_t>::max_items() const
 {
     return max_items_;
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 bool LruCache<K, T, mem, lookup_t>::has(const K& key) const
 {
     return lookup.find(key) != lookup.end();
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 const T& LruCache<K, T, mem, lookup_t>::at(const K& key)
 {
     auto it = lookup.at(key);
@@ -38,13 +38,13 @@ const T& LruCache<K, T, mem, lookup_t>::at(const K& key)
     return items.begin()->second;
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 const T& LruCache<K, T, mem, lookup_t>::operator[](const K& key)
 {
     return at(key);
 }
 
-template <typename K, typename T, size_t mem, typename lookup_t> 
+template <typename K, typename T, ull mem, typename lookup_t> 
 void LruCache<K, T, mem, lookup_t>::put(const K& key, const T& value)
 {
     items.push_front(std::make_pair(key, value));
